refactor(studio): split selected/ inspector and materials editor into focused modules
The inspector (`ui_panels/selected/mod.rs`) and the material editor
(`ui_panels/materials.rs`) had both grown past 1100 lines, mixing several
unrelated UI sections inside a single method. Extract the largest discrete
sections into sibling submodules so each file owns one concern:

- selected/mod.rs (1428 → 933) — extracted transform_grid.rs (gizmo-mode
  toggle + translate/rotate/scale grid), light_editor.rs (kind/colour/
  intensity + cone angles), modifiers.rs (CSG op switch + array/mirror
  rows). The wants_* deferred-action handlers stay in mod.rs since they
  share state across sections.
- materials.rs (1176) → materials/mod.rs (476) + pbr.rs (PBR slider/colour
  block) + textures.rs (thumbnail, slot list, manual slot picker — keeps
  thumb_handle / cached_exists alongside) + manage.rs (rename/delete chip
  plus next_material_name / rewrite_material_decl_name helpers).

Pure UI extraction — no behaviour change. mogen-studio tests still pass
(215/215).
